% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute2clu_function.R
\name{attribute2clu}
\alias{attribute2clu}
\title{attribute2clu}
\usage{
attribute2clu(df, attribute, NAME)
}
\arguments{
\item{df}{Attribute Dataframe}

\item{attribute}{Attribute Name - Column name (character) in dataframe to write to .clu file}

\item{NAME}{Name for .clu file}
}
\value{
.clu file
}
\description{
This function converts a attribute from a  dataframe to a .net files. Dataframe has two columns for binary, and three for weighted (third column is weights)
}
