% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute2vec_function.R
\name{attribute2vec}
\alias{attribute2vec}
\title{attribute2vec}
\usage{
attribute2vec(df, attribute, NAME)
}
\arguments{
\item{df}{Attribute Dataframe}

\item{attribute}{Attribute Name - Column name (character) in dataframe to write to .vec file}

\item{NAME}{Name for .vec file}
}
\value{
.vec file
}
\description{
This function converts attribute from a dataframe to a .vec file. Dataframe has two columns for binary, and three for weighted (third column is weights)
}
