% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist2pajek_function.R
\name{edgelist2pajek}
\alias{edgelist2pajek}
\title{edgelist2pajek}
\usage{
edgelist2pajek(df, weighted, directed, NAME)
}
\arguments{
\item{df}{Edgelist Dataframe}

\item{weighted}{TRUE/FALSE}

\item{directed}{TRUE/FALSE}

\item{NAME}{Name for .net file}
}
\value{
.net file
}
\description{
This function converts a edgelist dataframe to a .net files. Dataframe has two columns for binary, and three for weighted (third column is weights)
}
