% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2pajek_function.R
\name{matrix2pajek}
\alias{matrix2pajek}
\title{matrix2pajek}
\usage{
matrix2pajek(mat, weighted, directed, NAME)
}
\arguments{
\item{mat}{matrix}

\item{weighted}{TRUE/FALSE}

\item{directed}{TRUE/FALSE}

\item{NAME}{Name for .net file}
}
\value{
.net file
}
\description{
This function converts a matrix to a .net files. Dataframe has two columns for binary, and three for weighted (third column is weights)
}
